requirements.txt
# ── LLM & Agent Orchestration ─────────────────────────────────────────────────
langchain-groq==0.3.2
langgraph==1.0.8
langgraph-checkpoint==4.0.0
langgraph-prebuilt==1.0.7
langgraph-sdk==0.3.4

# ── Data Validation ────────────────────────────────────────────────────────────
pydantic==2.11.3

# ── News Fetching & Parsing ────────────────────────────────────────────────────
feedparser==6.0.11
beautifulsoup4==4.13.4

# ── Search & Contact Finding ───────────────────────────────────────────────────
ddgs==9.11.12
requests==2.32.3

# ── Fuzzy Deduplication ────────────────────────────────────────────────────────
rapidfuzz==3.14.3

# ── Data & Excel ──────────────────────────────────────────────────────────────
pandas==2.2.3
openpyxl==3.1.5
numpy==1.26.4

# ── Dashboard ─────────────────────────────────────────────────────────────────
streamlit==1.45.0

# ── Scheduler ─────────────────────────────────────────────────────────────────
apscheduler==3.11.0

# ── Environment & Config ──────────────────────────────────────────────────────
python-dotenv==1.1.0